refactor: step2 패키지에 대해서 네이밍 부여
