+++ src/test/java/stringcalculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package stringcalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
